+++ CHANGELOG.md
@@ -118,6 +118,9 @@ For instructions on upgrading to newer versions, visit
 * \#1381, \#1371 The identity map now functions properly with inherited
   documents. (Paul Canavese)
 
+* \#1370 Split concat on embedded arrays into its own method to handle the
+  batch processing due to after callback run execution issues.
+
 * \#1366 Array and hash values now get deep copied for dirty tracking.
 
 * \#1359 Provide ability to not have default scope applied to all named